CURD Operation can be done using only Verified User

+++ main.py
@@ -1,4 +1,6 @@
-from os import stat
+
+from pyexpat import model
+from tokenize import TokenInfo
 from fastapi import HTTPException
 from fastapi import FastAPI
 from fastapi import Depends
@@ -7,6 +9,7 @@ from pydantic import BaseModel,Field
 from sqlalchemy.orm import Session
 import models
 from database import SessionLocal, engine
+from auth import get_current_user ,get_user_exception
 
 app=FastAPI()
 
@@ -22,9 +25,8 @@ def get_db():
 class Todo(BaseModel):
     title:str
     description:Optional[str]
-    priority:int=Field(gt=0,lt=100,description="The priority must be between 1-10")
-    complete:bool
-
+    priority:int
+    complete:bool 
 
 
 @app.get("/")
@@ -32,35 +34,56 @@ async def read_all_todos(db:Session=Depends(get_db)):
     return db.query(models.Todos).all()
 
 @app.get("/todo/{todoid}")
-async def get_todo_by_ID(todo_id : int,db:Session=Depends(get_db)):
+async def get_todo_by_ID(todoid : int,user:dict=Depends(get_current_user),db:Session=Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+
     todo_model=db.query(models.Todos)\
-        .filter(models.Todos.id==todo_id)\
-        .first()
+        .filter(models.Todos.id==todoid)\
+        .filter(models.Todos.owner_id==user.get("id"))\
+        .all()
     if todo_model is not None:
         return todo_model
     raise http_exception()
-    
+
+@app.get("/todos/user")
+async def read_all_by_user(user:dict=Depends(get_current_user) , db:Session=Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+     
+    return db.query(models.Todos).filter(models.Todos.owner_id==user.get("id")).all()
+
+
 @app.post("/")
-async def create_todo(todo:Todo,db:Session=Depends(get_db)):
+async def create_todo(todo:Todo,user:dict=Depends(get_current_user),db:Session=Depends(get_db)):
+    if user is None:
+        raise  get_user_exception()
+    print(todo)
     todo_model=models.Todos()
     todo_model.title=todo.title
-    todo_model.desciption=todo.description
+    todo_model.description=todo.description
     todo_model.priority=todo.priority
     todo_model.complete=todo.complete
-
+    todo_model.owner_id=user.get('id')
+    
     db.add(todo_model)
     db.commit()
 
-    return {
-        "status":201,
-        'transaction':'successfull'
-    }
+    return success_response(200)
 
 @app.put("/{todo_id}")
-async def update_todo(todo_id:int,todo:Todo,db:Session=Depends(get_db)):
+async def update_todo(todo_id:int,todo:Todo, 
+                      user:dict=Depends(get_current_user),
+                      db:Session=Depends(get_db)):
+
+    if user is None:
+        raise get_user_exception()
+        
     todo_model=db.query(models.Todos)\
         .filter(models.Todos.id==todo_id)\
+        .filter(models.Todos.owner_id==user.get('id'))\
         .first()
+
 
     if todo_model is None:
         raise http_exception()
@@ -69,22 +92,26 @@ async def update_todo(todo_id:int,todo:Todo,db:Session=Depends(get_db)):
     todo_model.desciption=todo.description
     todo_model.priority=todo.priority
     todo_model.complete=todo.complete
+    
 
     db.add(todo_model)
     db.commit()
-    return {
-        "status":201,
-        'transaction':'successfull'
-    }
+    return success_response(200)
 
 @app.delete("/{todo_id}")
-async def delete_todo(todo_id:int, db:Session=Depends(get_db)):
-    todo_model=db.query(models.Todos).filter(models.Todos.id==todo_id).first()
+async def delete_todo(todo_id:int,
+                    user:dict=Depends(get_current_user),
+                    db:Session=Depends(get_db)):
+                    
+    if user is None:
+        raise get_user_exception()                
+    todo_model=db.query(models.Todos).filter(models.Todos.id==todo_id).filter(models.Todos.owner_id==user.get("id")).all()
     if todo_model is None:
         raise http_exception()
     db.query(models.Todos).filter(models.Todos.id==todo_id).delete()
     db.commit()
     return success_response(200)
+
 
 def success_response(status_code:int):
     return {
